[ALPS06427010] Extractor Fuzzer: fix asf-overflow-2 and coverity issue
fix 2 overflow issues in asf
fix coverity issue to avoid null dereference in asf

CR-Id: ALPS06427010
Feature: Vulnerability Scan

+++ ASFExtractor.h
@@ -134,7 +134,7 @@ public:
 
 protected:
     bool ParseASF();
-    int64_t ASFSeekTo(uint32_t seekTimeMs);
+    int64_t ASFSeekTo(int64_t seekTimeMs);
 
     bool RetrieveWmvCodecSpecificData(asf_stream_t *pStreamProp, AMediaFormat *meta,
             VC1SeqData *pCodecSpecificData);
